Add code to try removing problematic records before moving problematic files to their own directory.

+++ site-linux/convert_and_restructure_daily_on_nas.sh
@@ -27,6 +27,7 @@ RAWACF_ARRAY_DEST=/borealis_nfs/borealis_data/rawacf_array
 BFIQ_ARRAY_DEST=/borealis_nfs/borealis_data/bfiq_array
 ANTENNAS_IQ_ARRAY_DEST=/borealis_nfs/borealis_data/antennas_iq_array
 BACKUP_DEST=/borealis_nfs/borealis_data/backup
+PROBLEM_FILES_DEST=/borealis_nfs/borealis_data/conversion_failure
 
 LOGGINGDIR=${HOME}/logs/file_conversions/${CURYEAR}/${CURMONTH}
 mkdir -p ${LOGGINGDIR}
@@ -88,24 +89,32 @@ do
     echo "python3 ${HOME}/data_flow/site-linux/borealis_convert_file.py ${f}" >> ${LOGFILE} 2>&1
     python3 ${HOME}/data_flow/site-linux/borealis_convert_file.py ${f} >> ${LOGFILE} 2>&1
     ret=$?
-    if [ $ret -eq 0 ]; then
-        # move the resulting files if all was successful
-        # then remove the source site file.
-        dmap_file_start="${f%.rawacf.hdf5.site}"
+    if [ ! $ret -eq 0 ]; then
+        # Try to remove records
+        remove_record_output=$(python3 ${HOME}/data_flow/site-linux/remove_record.py ${f})
+        echo "$remove_record_output" >> ${LOGFILE} 2>&1
+        EMAILBODY="${EMAILBODY}\nAttempting to remove records from ${f}:\n${remove_record_output}"
+        ret=$?
+        if [ $ret -eq 0 ]; then
+            # move the resulting files if all was successful
+            # then remove the source site file.
+            dmap_file_start="${f%.rawacf.hdf5.site}"
 
-        # remove last character(s) (slice_id)
-        slice_id=${dmap_file_start##*.}
-        dmap_file_wo_slice_id=${dmap_file_start%${slice_id}}
+            # remove last character(s) (slice_id)
+            slice_id=${dmap_file_start##*.}
+            dmap_file_wo_slice_id=${dmap_file_start%${slice_id}}
 
-        ordinal_id="$(($slice_id + 97))"
-        file_character=`chr $ordinal_id`
-        dmap_file="${dmap_file_wo_slice_id}${file_character}.rawacf.bz2"
-        mv -v ${dmap_file} ${DMAP_DEST}/ >> ${LOGFILE} 2>&1
-        array_file="${f%.site}"
-        mv -v ${array_file} ${RAWACF_ARRAY_DEST} >> ${LOGFILE} 2>&1
-        rm -v ${f} >> ${LOGFILE} 2>&1
-    else
-        EMAILBODY="${EMAILBODY}\nFile failed to convert: ${f}"
+            ordinal_id="$(($slice_id + 97))"
+            file_character=`chr $ordinal_id`
+            dmap_file="${dmap_file_wo_slice_id}${file_character}.rawacf.bz2"
+            mv -v ${dmap_file} ${DMAP_DEST}/ >> ${LOGFILE} 2>&1
+            array_file="${f%.site}"
+            mv -v ${array_file} ${RAWACF_ARRAY_DEST} >> ${LOGFILE} 2>&1
+            rm -v ${f} >> ${LOGFILE} 2>&1
+        else
+            EMAILBODY="${EMAILBODY}\nFile failed to convert: ${f}"
+            mv -v ${f} ${PROBLEM_FILES_DEST} >> ${LOGFILE} 2>&1
+        fi
     fi
 done
 
@@ -115,25 +124,33 @@ do
     echo "python3 ${HOME}/data_flow/site-linux/borealis_convert_file.py ${f}" >> ${LOGFILE} 2>&1
     python3 ${HOME}/data_flow/site-linux/borealis_convert_file.py ${f} >> ${LOGFILE} 2>&1
     ret=$?
-    if [ $ret -eq 0 ]; then
-        # remove iqdat and move bfiq array file if successful.
-        # then remove source site file.
-        dmap_file_start="${f%.bfiq.hdf5.site}"
+    if [ ! $ret -eq 0 ]; then
+        # Try to remove records
+        remove_record_output=$(python3 ${HOME}/data_flow/site-linux/remove_record.py ${f})
+        echo "$remove_record_output" >> ${LOGFILE} 2>&1
+        EMAILBODY="${EMAILBODY}\nAttempting to remove records from ${f}:\n${remove_record_output}"
+        ret=$?
+        if [ $ret -eq 0 ]; then
+            # remove iqdat and move bfiq array file if successful.
+            # then remove source site file.
+            dmap_file_start="${f%.bfiq.hdf5.site}"
 
-        # remove last character(s) (slice_id)
-        slice_id=${dmap_file_start##*.}
-        dmap_file_wo_slice_id=${dmap_file_start%${slice_id}}
+            # remove last character(s) (slice_id)
+            slice_id=${dmap_file_start##*.}
+            dmap_file_wo_slice_id=${dmap_file_start%${slice_id}}
 
-        ordinal_id="$(($slice_id + 97))"
-        file_character=`chr $ordinal_id`
-        dmap_file="${dmap_file_wo_slice_id}${file_character}.iqdat.bz2"
+            ordinal_id="$(($slice_id + 97))"
+            file_character=`chr $ordinal_id`
+            dmap_file="${dmap_file_wo_slice_id}${file_character}.iqdat.bz2"
 
-        rm -v ${dmap_file} >> ${LOGFILE} 2>&1
-        array_file="${f%.site}"
-        mv -v ${array_file} ${BFIQ_ARRAY_DEST} >> ${LOGFILE} 2>&1
-        rm -v ${f} >> ${LOGFILE} 2>&1
-    else
-        EMAILBODY="${EMAILBODY}\nFile failed to convert: ${f}"
+            rm -v ${dmap_file} >> ${LOGFILE} 2>&1
+            array_file="${f%.site}"
+            mv -v ${array_file} ${BFIQ_ARRAY_DEST} >> ${LOGFILE} 2>&1
+            rm -v ${f} >> ${LOGFILE} 2>&1
+        else
+            EMAILBODY="${EMAILBODY}\nFile failed to convert: ${f}"
+            mv -v ${f} ${PROBLEM_FILES_DEST} >> ${LOGFILE} 2>&1
+        fi
     fi
 done
 
@@ -147,14 +164,22 @@ else
     echo "python3 ${HOME}/data_flow/site-linux/borealis_convert_file.py ${f}" >> ${LOGFILE} 2>&1
     python3 ${HOME}/data_flow/site-linux/borealis_convert_file.py ${f} >> ${LOGFILE} 2>&1
     ret=$?
-    if [ $ret -eq 0 ]; then
-        # remove iqdat and move bfiq array file if successful.
-        # then remove source site file.
-        array_file="${f%.site}"
-        mv -v ${array_file} ${ANTENNAS_IQ_ARRAY_DEST} >> ${LOGFILE} 2>&1
-        rm -v ${f} >> ${LOGFILE} 2>&1
-    else
-        EMAILBODY="${EMAILBODY}\nFile failed to convert: ${f}"
+    if [ ! $ret -eq 0 ]; then
+        # Try to remove records
+        remove_record_output=$(python3 ${HOME}/data_flow/site-linux/remove_record.py ${f})
+        echo "$remove_record_output" >> ${LOGFILE} 2>&1
+        EMAILBODY="${EMAILBODY}\nAttempting to remove records from ${f}:\n${remove_record_output}"
+        ret=$?
+        if [ $ret -eq 0 ]; then
+            # remove iqdat and move bfiq array file if successful.
+            # then remove source site file.
+            array_file="${f%.site}"
+            mv -v ${array_file} ${ANTENNAS_IQ_ARRAY_DEST} >> ${LOGFILE} 2>&1
+            rm -v ${f} >> ${LOGFILE} 2>&1
+        else
+            EMAILBODY="${EMAILBODY}\nFile failed to convert: ${f}"
+            mv -v ${f} ${PROBLEM_FILES_DEST} >> ${LOGFILE} 2>&1
+        fi
     fi
 fi
 done
